REFACTOR: folder renamed Accessibility
	componentsAccessibility renamed to AccessibilityScripts
