+++ accessibilityScripts/focusWindowOnClick.ts
