+++ accessibilityScripts/trapFocus.ts
