+++ accessibilityScripts/windowChildrenTabIndexHandler.ts
